Course Timetabling: XML loader  - added ability to skip purging invalid placements phase (when solver parameter General.PurgeInvalidPlacements is set to false, defaults to true)

+++ src/net/sf/cpsolver/coursett/TimetableXMLLoader.java
@@ -744,10 +744,12 @@ public class TimetableXMLLoader extends TimetableLoader {
             }
         }
 
-        iProgress.setPhase("Purging invalid placements ...", getModel().variables().size());
-        for (Lecture lecture : getModel().variables()) {
-            lecture.purgeInvalidValues(iInteractiveMode);
-            iProgress.incProgress();
+        if (getModel().getProperties().getPropertyBoolean("General.PurgeInvalidPlacements", true)) {
+            iProgress.setPhase("Purging invalid placements ...", getModel().variables().size());
+            for (Lecture lecture : getModel().variables()) {
+                lecture.purgeInvalidValues(iInteractiveMode);
+                iProgress.incProgress();
+            }            
         }
         
         if (getModel().hasConstantVariables() && getModel().constantVariables().size() > 0) {
